Ignore action derivation of overlay elements by a modal
- Try to solve issue #230
- Try to automatically detect the high z-index modal element of the web page, to only derive actions on its childrens.
- Probably, an implementation that allows customizing modal elements using "class" web property it could be more precise

- Take web z-index property of all widgets
- Check if some web element is a modal (display = "block")
- If exists some widget that is a modal with a high web z-index, mark as widgetModal
- Create helper functionality "widgetIsChildOfParent"
- Improve web action descriptions

+++ webdriver/src/org/fruit/alayer/webdriver/enums/WdTags.java
@@ -134,7 +134,7 @@ public class WdTags extends TagsBase {
   /**
    * Web zIndex of a widget
    */
-  public static final Tag<Double> WebZIndex = from("WebZIndex", Double.class);
+  public static final Tag<Integer> WebZIndex = from("WebZIndex", Integer.class);
 
   /**
    * Web Control Type Identifier of a widget.
